converse: Write to stdout during timeout
PAM will not pass unknown error codes returned by the conversation
function back to the application (see pam_conv(3) "On failure, the
conversation function should ... return one of the predefined PAM error
codes").

Since there is no predefined timeout error code, we're left with either
reporting the timeout on stdout/stderr from within the context of the
conversation function (as is done in og-sudo) or not reporting it at all.